+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 
 v0.9.24
 =======
-* Added XPath 1.0/2.0 full parsing with elementpath
+* Added XPath 1.0/2.0 full parsing with the derived *elementpath* package
 * Test package improved (tox.ini, other checks with test_package.py)
 
 v0.9.23